Add R2GS baseline experiment results (9/15 complete)
Results for head, abdomen, pancreas (3/6/9 views):
- head:     26.87 / 32.87 / 35.48 PSNR
- abdomen:  29.88 / 34.43 / 36.71 PSNR
- pancreas: -     / 33.79 / 35.53 PSNR

chest, foot, pancreas_3views still running.
Fix float64 data type issue in dataset_readers.py.

+++ r2_gaussian/dataset/dataset_readers.py
@@ -267,7 +267,7 @@ def readNAFInfo(path, eval):
             )  # R is stored transposed due to 'glm' in CUDA code
             T = w2c[:3, 3]
 
-            image = projs[i_split] * scene_scale
+            image = np.float32(projs[i_split]) * scene_scale
 
             # Note, dDetector is [v, u] not [u, v]
             FovX = np.arctan2(scanner_cfg["sDetector"][1] / 2, scanner_cfg["DSD"]) * 2
